pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ragpipeline"
version = "0.0.1"
description = "An Opensource Implementation of RAG LLM Pipeline."
authors = [
  { name = "VenkateshTantravahi" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    # core utils
    "ipython",
    "jupyterlab",
    "loguru",
    "matplotlib",
    "notebook",
    "numpy",
    "pandas",
    "pip",
    "pytest",
    "python-dotenv",
    "ruff",
    "tqdm",
    "typer",

    # RAG & AI Stack
    "langchain",
    "langchain-community",
    "langchain-chroma",
    "langchain-huggingface",
    "langchain-text-splitters",
    "marker-pdf",
    "chromadb",
    "sentence-transformers",
    "ollama",
    "pillow-heif"
]
requires-python = "~=3.12.0"


[tool.ruff]
line-length = 99
src = ["RAGPipeline"]
include = ["pyproject.toml", "RAGPipeline/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["RAGPipeline"]
force-sort-within-sections = true


